Refactor pipeline config to separate file
