+++ .cursorignore
@@ -1,4 +1,6 @@
 # Add directories or file patterns to ignore during indexing (e.g. foo/ or *.csv)
-data
-profiles
-drizzle/meta
+data/
+profiles/
+drizzle/meta/
+# do not ignore app
+!src/**/*/
